Made index into barrel and made seperate drawing file
